requirements.txt
fastapi
uvicorn
langchain
openai
langchain-community
langchain-openai
PyPDF2
python-multipart

tavily-python